+++ src/engine/core/tests/dubeolsik.rs
@@ -52,6 +52,16 @@ fn esc() {
     ]);
 }
 
+// issue #410
+#[test]
+fn other_keys() {
+    let pass_keys = [Esc, PageUp, PageDown, Home, End];
+
+    for key in pass_keys.iter().copied() {
+        test_input(&[(Key::normal(R), "ㄱ", ""), (Key::normal(key), "", "ㄱPASS")]);
+    }
+}
+
 // issue #373
 #[test]
 fn arrow() {