+++ docs/CHANGELOG.md
@@ -6,6 +6,8 @@
 
 ### Improve
 
+* Fix Home, End, PageUp, PageDown don't clear preedit [#410](https://github.com/Riey/kime/issues/410)
+
 ## 2.2.0
 
 ### Improve